requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
dnspython==2.5.0
email-validator==2.1.0.post1
Faker==22.5.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
